+++ src/system/esp_ll_template.c
@@ -43,11 +43,9 @@ static uint8_t initialized = 0;
  * \param[in]       len: Number of bytes to send
  * \return          Number of bytes sent
  */
-static uint16_t
-send_data(const void* data, uint16_t len) {
-    /*
-     * Implement send function here
-     */
+static size_t
+send_data(const void* data, size_t len) {
+    /* Implement send function here */
     
     
     return len;                                 /* Return number of bytes actually sent to AT port */
